third_party/WebKit/LayoutTests/inspector/console/console-timeline-expected.txt

CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'one' finished.
CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'two' started.
CONSOLE DEBUG: Timeline 'two' started.
CONSOLE DEBUG: Timeline 'two' finished.
CONSOLE DEBUG: Timeline 'one' finished.
CONSOLE DEBUG: Timeline 'two' finished.
CONSOLE DEBUG: Timeline 'one' finished.
CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'two' started.
CONSOLE DEBUG: Timeline 'two' finished.
CONSOLE DEBUG: Timeline 'one' finished.
CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'two' was not started.
CONSOLE DEBUG: Timeline 'one' finished.
CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'one' finished.
CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'two' started.
CONSOLE DEBUG: Timeline 'one' terminated.
CONSOLE DEBUG: Timeline 'two' terminated.
CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'two' started.
CONSOLE DEBUG: Timeline 'one' terminated.
CONSOLE DEBUG: Timeline 'two' terminated.
CONSOLE DEBUG: Timeline 'one' started.
CONSOLE DEBUG: Timeline 'two' started.
CONSOLE DEBUG: Timeline 'one' terminated.
CONSOLE DEBUG: Timeline 'two' terminated.
Tests console.timeline and timelineEnd commands.


Running: testStartStopTimeline
Timeline 'one' started.
timestamp 1
Timeline 'one' finished.

Running: testStartStopMultiple
Timeline 'one' started.
timestamp 1
Timeline 'one' started.
Timeline 'two' started.
Timeline 'two' started.
Timeline 'two' finished.
timestamp 2
Timeline 'one' finished.
timestamp 3
Timeline 'two' finished.
timestamp 4
Timeline 'one' finished.

Running: testStartMultipleStopInsideEvals
Timeline 'one' started.
timestamp 1
Timeline 'two' started.
timestamp 2
timestamp 3
Timeline 'two' finished.
timestamp 4
timestamp 5
Timeline 'one' finished.

Running: testStopUnknown
Timeline 'one' started.
timestamp 1
timestamp 2
Timeline 'one' finished.

Running: testStartFromPanel
timestamp 0
Timeline 'one' started.
timestamp 1
Timeline 'one' finished.
timestamp 2

Running: testStopFromPanel
Timeline 'one' started.
timestamp 1
Timeline 'two' started.
timestamp 2

Running: testRacyStart
Timeline 'one' started.
timestamp 1
Timeline 'two' started.
timestamp 2

Running: testRacyStart2
timestamp 0
Timeline 'one' started.
timestamp 1
Timeline 'two' started.
timestamp 2
